+++ internal/dataplane/translator/translate_certs_test.go
@@ -3,6 +3,7 @@ package translator
 import (
 	"sort"
 	"testing"
+	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/kong/go-kong/kong"
@@ -408,6 +409,51 @@ func TestMergeCerts(t *testing.T) {
 			},
 		},
 		{
+			name: "multiple certs with same content should be merged, and tags should inherit from the cert with the earliest CreationTimestamp",
+			certs: []certWrapper{
+				{
+					identifier: string(crt1) + string(key1),
+					cert: kong.Certificate{
+						ID:   kong.String("certificate-1"),
+						Cert: kong.String(string(crt1)),
+						Key:  kong.String(string(key1)),
+						Tags: kong.StringSlice("tag1", "tag2"),
+					},
+					snis:              []string{"foo.com"},
+					CreationTimestamp: metav1.NewTime(time.Now()),
+				},
+				{
+					identifier: string(crt1) + string(key1),
+					cert: kong.Certificate{
+						ID:   kong.String("certificate-1-1"),
+						Cert: kong.String(string(crt1)),
+						Key:  kong.String(string(key1)),
+						Tags: kong.StringSlice("tag3", "tag4"),
+					},
+					snis:              []string{"baz.com"},
+					CreationTimestamp: metav1.NewTime(time.Now().Add(-1 * time.Hour)),
+				},
+			},
+			mergedCerts: []kongstate.Certificate{
+				{
+					Certificate: kong.Certificate{
+						ID:   kong.String("certificate-1-1"),
+						Cert: kong.String(string(crt1)),
+						Key:  kong.String(string(key1)),
+						// SNIs should be sorted
+						SNIs: kong.StringSlice("baz.com", "foo.com"),
+						// tags should inherit from the cert with the earliest CreationTimestamp.
+						Tags: kong.StringSlice("tag3", "tag4"),
+					},
+				},
+			},
+			idToMergedID: certIDToMergedCertID{
+				// the cert with the earliest CreationTimestamp should be the merged cert ID
+				"certificate-1":   "certificate-1-1",
+				"certificate-1-1": "certificate-1-1",
+			},
+		},
+		{
 			name: "multiple certs with same content should be merged, and tags should inherit from the cert with the lowest ID",
 			certs: []certWrapper{
 				{
@@ -449,7 +495,7 @@ func TestMergeCerts(t *testing.T) {
 						Key:  kong.String(string(key1)),
 						// SNIs should be sorted
 						SNIs: kong.StringSlice("baz.com", "foo.com"),
-						// tags should inherit from the first cert
+						// tags should inherit from the the cert with the lowest ID.
 						Tags: kong.StringSlice("tag1", "tag2"),
 					},
 				},